v1

+++ hb-telegram-bot/src/main/java/mavmi/telegram_bot/hb/sheets/dto/CellPosition.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class CellPosition {
 
     private List<Character> column;
-    private int row;
+    private long row;
 
     public void moveUp() {
         row--;
